Fix validation method names in B2C, balance, and C2B packages
The validation method names in the B2C, balance, and C2B packages were incorrectly capitalized. This commit fixes the method names to match the correct capitalization.
