+++ pkg/request.go
@@ -1,45 +1,5 @@
 package pkg
 
-import (
-	"errors"
-	"net/url"
-)
-
-var (
-	// errInvalidCommandID indicates the CommandID is invalid.
-	errInvalidCommandID = errors.New("invalid command id")
-
-	// errInvalidTransactionType indicates invalid transaction type.
-	errInvalidTransactionType = errors.New("invalid transaction type")
-
-	// errInvalidPhoneNumber indicates invalid phone number.
-	errInvalidPhoneNumber = errors.New("invalid phone number")
-
-	// errInvalidShortCode indicates invalid short code.
-	errInvalidShortCode = errors.New("invalid short code")
-
-	// errInvalidAccountReference indicates invalid account reference.
-	errInvalidAccountReference = errors.New("invalid account reference")
-
-	// errInvalidTransactionDesc indicates invalid transaction description.
-	errInvalidTransactionDesc = errors.New("invalid transaction description")
-
-	// errInvalidRemarks indicates invalid remarks.
-	errInvalidRemarks = errors.New("invalid remarks")
-
-	// errInvalidOccasion indicates invalid occasion.
-	errInvalidOccasion = errors.New("invalid occasion")
-
-	// errInvalidResponseType indicates invalid response type.
-	errInvalidResponseType = errors.New("invalid response type")
-
-	// errInvalidIdentifierType indicates invalid identifier type.
-	errInvalidIdentifierType = errors.New("invalid identifier type")
-
-	// errInvalidURL indicates invalid url.
-	errInvalidURL = errors.New("invalid url")
-)
-
 // ExpressSimulateReq struct.
 type ExpressSimulateReq struct {
 	PassKey           string `json:"PassKey,omitempty"`           // This is a passkey which is generated by Safaricom for use when making Lipa Na M-Pesa Online Payment API requests.
@@ -56,42 +16,6 @@ type ExpressSimulateReq struct {
 	TransactionDesc   string `json:"TransactionDesc,omitempty"`   // This is any additional information/comment that can be sent along with the request
 }
 
-// Validate validate the ExpressSimulateReq Request.
-func (esr ExpressSimulateReq) validate() error {
-	if ok := isShortCode(esr.BusinessShortCode); !ok {
-		return errInvalidShortCode
-	}
-
-	if esr.TransactionType != "CustomerPayBillOnline" && esr.TransactionType != "CustomerBuyGoodsOnline" {
-		return errInvalidTransactionType
-	}
-
-	if ok := isPhoneNumber(esr.PartyA); !ok {
-		return errInvalidPhoneNumber
-	}
-
-	if ok := isShortCode(esr.PartyB); !ok {
-		return errInvalidShortCode
-	}
-
-	if ok := isPhoneNumber(esr.PhoneNumber); !ok {
-		return errInvalidPhoneNumber
-	}
-
-	if len(esr.AccountReference) > 12 {
-		return errInvalidAccountReference
-	}
-
-	if len(esr.TransactionDesc) > 13 {
-		return errInvalidTransactionDesc
-	}
-	if ok := isValidURL(esr.CallBackURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
-}
-
 // ExpressQueryReq struct.
 type ExpressQueryReq struct {
 	PassKey           string `json:"PassKey,omitempty"`           // This is a passkey which is generated by Safaricom for use when making Lipa Na M-Pesa Online Payment API requests.
@@ -99,15 +23,6 @@ type ExpressQueryReq struct {
 	Password          string `json:"Password,omitempty"`          // This is the password used for encrypting the request sent: A base64 encoded string.
 	Timestamp         string `json:"Timestamp,omitempty"`         // This is the Timestamp of the transaction, normally in the formart of YEAR+MONTH+DATE+HOUR+MINUTE+SECOND (YYYYMMDDHHMMSS)
 	CheckoutRequestID string `json:"CheckoutRequestID,omitempty"` // This is a global unique identifier of the processed checkout transaction request.
-}
-
-// Validate validate the ExpressQueryReq Request.
-func (eqr ExpressQueryReq) validate() error {
-	if ok := isShortCode(eqr.BusinessShortCode); !ok {
-		return errInvalidShortCode
-	}
-
-	return nil
 }
 
 // GenerateQRReq struct.
@@ -120,38 +35,12 @@ type GenerateQRReq struct {
 	Size         string `json:"Size,omitempty"`         // Size of the QR code image in pixels. QR code image will always be a square image.
 }
 
-// Validate validate the GenerateQRReq Request.
-func (qr GenerateQRReq) validate() error {
-	if qr.TrxCode != "SB" && qr.TrxCode != "SM" && qr.TrxCode != "PB" && qr.TrxCode != "WA" && qr.TrxCode != "BG" {
-		return errInvalidTransactionType
-	}
-
-	return nil
-}
-
 // C2BRegisterURLReq struct.
 type C2BRegisterURLReq struct {
 	ValidationURL   string `json:"ValidationURL,omitempty"`   // This is the URL that receives the validation request from API upon payment submission. The validation URL is only called if external validation on the registered shortcode is enabled. (By default external validation is disabled)
 	ConfirmationURL string `json:"ConfirmationURL,omitempty"` // Thie is the URL that receives the confirmation request from API upon payment completion
 	ShortCode       uint64 `json:"ShortCode,omitempty"`       // The shortcode of the organization
 	ResponseType    string `json:"ResponseType,omitempty"`    // This parameter specifies what is to happen if for any reason the validation URL is not reachable. It can be cancelled or completed.
-}
-
-func (c2b C2BRegisterURLReq) validate() error {
-	if ok := isShortCode(c2b.ShortCode); !ok {
-		return errInvalidShortCode
-	}
-	if c2b.ResponseType != "Completed" && c2b.ResponseType != "Cancelled" {
-		return errInvalidResponseType
-	}
-	if ok := isValidURL(c2b.ValidationURL); !ok {
-		return errInvalidURL
-	}
-	if ok := isValidURL(c2b.ConfirmationURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
 }
 
 // C2BSimulateReq struct.
@@ -161,14 +50,6 @@ type C2BSimulateReq struct {
 	BillRefNumber string `json:"BillRefNumber,omitempty"` // This is used on CustomerPayBillOnline option only. This is where a customer is expected to enter a unique bill identifier, e.g. an Account Number.
 	Amount        uint64 `json:"Amount,omitempty"`        // This is the amount being transacted.
 	ShortCode     uint64 `json:"ShortCode,omitempty"`     // This is the Short Code receiving the amount being transacted.
-}
-
-func (c2b C2BSimulateReq) validate() error {
-	if c2b.CommandID != "CustomerPayBillOnline" && c2b.CommandID != "CustomerBuyGoodsOnline" {
-		return errInvalidCommandID
-	}
-
-	return nil
 }
 
 // B2CPaymentReq struct.
@@ -188,29 +69,6 @@ type B2CPaymentReq struct {
 	Amount                   uint64 `json:"Amount,omitempty"`   // The amount of money being sent to the customer.
 }
 
-// Validate validate the struct.
-func (r B2CPaymentReq) validate() error {
-	if r.CommandID != "BusinessPayment" && r.CommandID != "SalaryPayment" && r.CommandID != "PromotionPayment" {
-		return errInvalidCommandID
-	}
-
-	if ok := isShortCode(r.PartyA); !ok {
-		return errInvalidShortCode
-	}
-
-	if ok := isPhoneNumber(r.PartyB); !ok {
-		return errInvalidPhoneNumber
-	}
-	if ok := isValidURL(r.QueueTimeOutURL); !ok {
-		return errInvalidURL
-	}
-	if ok := isValidURL(r.ResultURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
-}
-
 // TransactionStatusReq struct.
 type TransactionStatusReq struct {
 	CommandID          string `json:"CommandID,omitempty"`          // Takes only the 'TransactionStatusQuery' Command ID.
@@ -226,35 +84,6 @@ type TransactionStatusReq struct {
 	Occasion           string `json:"Occasion,omitempty"`
 }
 
-// Validate validate the struct.
-func (r TransactionStatusReq) validate() error {
-	if r.CommandID != "TransactionStatusQuery" {
-		return errInvalidCommandID
-	}
-
-	if len(r.Remarks) > 100 {
-		return errInvalidRemarks
-	}
-
-	if r.Occasion != "" && len(r.Occasion) > 100 {
-		return errInvalidOccasion
-	}
-
-	if r.IdentifierType != 1 && r.IdentifierType != 2 && r.IdentifierType != 4 {
-		return errInvalidIdentifierType
-	}
-
-	if ok := isValidURL(r.QueueTimeOutURL); !ok {
-		return errInvalidURL
-	}
-
-	if ok := isValidURL(r.ResultURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
-}
-
 // AccountBalanceReq struct.
 type AccountBalanceReq struct {
 	CommandID          string `json:"CommandID,omitempty"`
@@ -266,27 +95,6 @@ type AccountBalanceReq struct {
 	SecurityCredential string `json:"SecurityCredential,omitempty"` // Encrypted Credential of user getting transaction amoun
 	QueueTimeOutURL    string `json:"QueueTimeOutURL,omitempty"`    // The path that stores information of time out transaction
 	ResultURL          string `json:"ResultURL,omitempty"`          // The path that stores information of transaction
-}
-
-// Validate validate the struct.
-func (r AccountBalanceReq) validate() error {
-	if r.CommandID != "AccountBalance" {
-		return errInvalidCommandID
-	}
-
-	if r.IdentifierType != 1 && r.IdentifierType != 2 && r.IdentifierType != 4 {
-		return errInvalidIdentifierType
-	}
-
-	if ok := isValidURL(r.QueueTimeOutURL); !ok {
-		return errInvalidURL
-	}
-
-	if ok := isValidURL(r.ResultURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
 }
 
 // ReverseReq struct.
@@ -305,23 +113,6 @@ type ReverseReq struct {
 	TransactionID          string `json:"TransactionID,omitempty"`          // Organization Receiving the funds.
 }
 
-// Validate validate the struct.
-func (r ReverseReq) validate() error {
-	if r.CommandID != "TransactionReversal" {
-		return errInvalidCommandID
-	}
-
-	if ok := isValidURL(r.QueueTimeOutURL); !ok {
-		return errInvalidURL
-	}
-
-	if ok := isValidURL(r.ResultURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
-}
-
 // RemitTaxReq struct.
 type RemitTaxReq struct {
 	InitiatorName          string `json:"Initiator,omitempty"`              // The name of Initiator to initiating  the request
@@ -337,62 +128,4 @@ type RemitTaxReq struct {
 	QueueTimeOutURL        string `json:"QueueTimeOutURL,omitempty"`        // The path that stores information of time out transaction
 	ResultURL              string `json:"ResultURL,omitempty"`              // The path that stores information of transaction
 	Remarks                string `json:"Remarks,omitempty"`                // Comments that are sent along with the transaction.
-}
-
-// Validate validate the struct.
-func (r RemitTaxReq) validate() error {
-	if r.CommandID != "PayTaxToKRA" {
-		return errInvalidCommandID
-	}
-
-	if len(r.Remarks) > 100 {
-		return errInvalidRemarks
-	}
-
-	if ok := isValidURL(r.QueueTimeOutURL); !ok {
-		return errInvalidURL
-	}
-
-	if ok := isValidURL(r.ResultURL); !ok {
-		return errInvalidURL
-	}
-
-	return nil
-}
-
-// isPhoneNumber checks if the number is a valid phone number.
-// MSISDN (12 digits Mobile Number) e.g. 2547XXXXXXXX.
-func isPhoneNumber(number uint64) bool {
-	if number < 100000000000 || number > 999999999999 {
-		return false
-	}
-
-	return true
-}
-
-// isShortCode checks if the number is a valid short code.
-// Shortcode (5 to 7 digits) e.g. 654321.
-func isShortCode(number uint64) bool {
-	if number < 10000 || number > 9999999 {
-		return false
-	}
-
-	return true
-}
-
-func isValidURL(inputURL string) bool {
-	parsedURL, err := url.Parse(inputURL)
-	if err != nil {
-		return false
-	}
-
-	if parsedURL.Scheme != "http" && parsedURL.Scheme != "https" {
-		return false
-	}
-
-	if parsedURL.Host == "" {
-		return false
-	}
-
-	return true
 }
